refactor: 그룹 멤버 생성 및 추가 기능을 createGroupMembers로 통합
