+++ src/main/java/com/dnd/moddo/domain/groupMember/controller/GroupMemberController.java
@@ -3,7 +3,6 @@ package com.dnd.moddo.domain.groupMember.controller;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -26,29 +25,19 @@ public class GroupMemberController {
 
 	@PostMapping
 	public ResponseEntity<GroupMembersResponse> saveGroupMembers(
-		@RequestParam("groupId") String token,
+		@RequestParam("groupId") Long groupId,
 		@Valid @RequestBody GroupMembersSaveRequest request
 	) {
-		Long groupId = 1L; //mock value
-		GroupMembersResponse response = commandGroupMemberService.createGroupMembers(groupId, request);
+		GroupMembersResponse response = commandGroupMemberService.addOrCreateGroupMembers(groupId, request);
 		return ResponseEntity.ok(response);
 	}
 
 	@GetMapping
 	public ResponseEntity<GroupMembersResponse> getGroupMembers(
-		@RequestParam("groupId") String token
+		@RequestParam("groupId") Long groupId
 	) {
-		Long groupId = 1L;
 		GroupMembersResponse response = queryGroupMemberService.findAll(groupId);
 		return ResponseEntity.ok(response);
 	}
 
-	@PutMapping
-	public ResponseEntity<GroupMembersResponse> updateGroupMembers(
-		@RequestParam("groupId") String token, @Valid @RequestBody GroupMembersSaveRequest request
-	) {
-		Long groupId = 1L;
-		GroupMembersResponse response = commandGroupMemberService.addGroupMembers(groupId, request);
-		return ResponseEntity.ok(response);
-	}
 }
